+++ tools/CHANGELOG.md
@@ -5,6 +5,8 @@ This changelog summarizes major changes between Truffle Tools versions.
 ## Version 24.1.0
 * GR-52742: `CPUSampler` no longer guarantees to keep all contexts on the engine alive. As a result `CPUSampler#getData()` is deprecated and may not return data for all contexts on the engine. The contexts that were already collected by GC won't be in the returned map. `CPUSamplerData#getContext()` is also deprecated and returns null if the context was already collected.
 * GR-52742: `CPUSamplerData#getDataList()` was introduced and returns all data collected by the sampler as a list of `CPUSamplerData`. For each context on the engine, including the ones that were already collected, there is a corresponding element in the list. `CPUSamplerData#getContextIndex()` returns the index of the data in the list.
+* GR-53035: Added CPUSampler support for gathering any async stack trace information for each sample, enabled by default. Introduced configuration option `--cpusampler.GatherAsyncStackTrace=true|false`, and corresponding API [`CPUSampler#setGatherAsyncStackTrace`](https://www.graalvm.org/tools/javadoc/com/oracle/truffle/tools/profiler/CPUSampler.html#setGatherAsyncStackTrace(boolean)), to allow disabling async stack trace sampling again.
+* GR-53035: `StackTraceEntry#isInlined()` is no longer deprecated.
 
 ## Version 23.0.0
 * GR-41407: Added new option `--dap.SourcePath` to allow to resolve sources with relative paths.